+++ types/interfaces/AuthenticationOptions.d.ts
@@ -1,3 +1,4 @@
+import express from 'express';
 /**
  * User authentication configuration interface.
  */
@@ -48,20 +49,20 @@ export default interface  {
      * Note that the user information must include an ID value that can identify the user.
      *
      * @example
-     * authenticate_user: async (username, password) => {
+     * authenticate_user: async (username, password, req) => {
      *   const UserModel = require('../models/UserModel');
      *   return UserModel.findOne({
      *     where: {
      *       email: username,
-     *       password: password
+     *       password
      *     },
      *     raw: true
      *   });
      * }
      *
-     * @type {(username: string, password: string) => Promise<{[key: string]: any}|null>}
+     * @type {(username: string, password: string, req: express.Request) => Promise<{[key: string]: any}|null>}
      */
-    authenticate_user: (username: string, password: string) => Promise<{
+    authenticate_user: (username: string, password: string, req: express.Request) => Promise<{
         [key: string]: any;
     } | null>;
     /**